auto-improve: Extract agent-invocation concerns from subprocess_utils into cai_lib/subagent/
Closes #1230.

Splits the 964-line ``cai_lib/subprocess_utils.py`` into a dedicated
``cai_lib/subagent/`` package with one clear role per module:

- ``core``        — ``run_subagent`` + ``_collect_results`` + ``_CLI_PATH``
- ``legacy``      — deprecated ``_run_claude_p`` / ``_argv_to_options`` facade
- ``cost``        — cost-row split + ``_post_cost_comment``
- ``stderr_sink`` — ``_make_stderr_sink`` / ``_captured_stderr_text`` + bounds
- ``fsm_state``   — ``_CURRENT_FSM_STATE`` contextvar + ``set_current_fsm_state``
- ``errors``      — ``_sdk_error_summary``

``cai_lib/subprocess_utils.py`` shrinks to 28 lines — just the ``_run``
shell wrapper that the 26 importers still use.

Zero behaviour change: function bodies are copied verbatim, only
import paths rewire. All 26 importers and three test files
(``test_subprocess_utils``, ``test_cost_comment``, ``test_sdk_spike_parity``)
have their imports and ``patch.object`` targets updated.

``run_subagent`` / ``_run_claude_p`` / ``set_current_fsm_state`` remain
importable from ``cai_lib.subagent``; ``_run`` remains in
``cai_lib.subprocess_utils``.

Verification:
- ``grep def _run_claude_p|def run_subagent|def _argv_to_options|def _collect_results cai_lib/subprocess_utils.py`` → 0
- ``grep def _run_claude_p|def run_subagent cai_lib/subagent/`` → 1 each
- ``wc -l cai_lib/subprocess_utils.py`` → 28
- ``python -m unittest discover tests`` → 771 tests OK (1 skipped)
- Remaining ``from cai_lib.subprocess_utils import`` lines are all ``_run``
