+++ cai_lib/actions/maintain.py
@@ -31,7 +31,8 @@ from cai_lib.fsm import (
 )
 from cai_lib.github import _build_issue_block, close_issue_completed
 from cai_lib.logging_utils import log_run
-from cai_lib.subprocess_utils import _run, _run_claude_p
+from cai_lib.subagent import _run_claude_p
+from cai_lib.subprocess_utils import _run
 
 
 # Matches the marker emitted by cai-maintain when the op list was
