+++ cai_lib/actions/plan.py
@@ -45,7 +45,8 @@ from cai_lib.github import (
     _post_issue_comment,
     _set_labels,
 )
-from cai_lib.subprocess_utils import _run, _run_claude_p
+from cai_lib.subagent import _run_claude_p
+from cai_lib.subprocess_utils import _run
 from cai_lib.logging_utils import log_run
 from cai_lib.cmd_helpers import (
     _work_directory_block,
