+++ cai_lib/actions/refine.py
@@ -19,7 +19,8 @@ from cai_lib.config import (
 )
 from cai_lib.fsm import fire_trigger
 from cai_lib.github import _build_issue_block
-from cai_lib.subprocess_utils import _run, _run_claude_p
+from cai_lib.subagent import _run_claude_p
+from cai_lib.subprocess_utils import _run
 from cai_lib.logging_utils import log_run
 from cai_lib.cmd_helpers import _strip_stored_plan_block
 from cai_lib.issues import create_issue, get_parent_issue, link_sub_issue, list_sub_issues
