+++ cai_lib/actions/review_pr.py
@@ -43,7 +43,8 @@ from cai_lib.fsm import (
 )
 from cai_lib.github import _fetch_linked_issue_block
 from cai_lib.logging_utils import log_run
-from cai_lib.subprocess_utils import _run, _run_claude_p
+from cai_lib.subagent import _run_claude_p
+from cai_lib.subprocess_utils import _run
 
 
 # ---------------------------------------------------------------------------
