+++ cai_lib/audit/runner.py
@@ -28,7 +28,8 @@ from cai_lib.audit.cost import _build_cost_summary
 from cai_lib.audit_logging import audit_log_finish, audit_log_start
 from cai_lib.config import PUBLISH_SCRIPT
 from cai_lib.logging_utils import log_run
-from cai_lib.subprocess_utils import _run, _run_claude_p
+from cai_lib.subagent import _run_claude_p
+from cai_lib.subprocess_utils import _run
 
 
 MODULES_YAML_REL = Path("docs/modules.yaml")
