+++ cai_lib/cmd_rescue.py
@@ -50,7 +50,7 @@ from cai_lib.github import (
     open_blockers,
 )
 from cai_lib.logging_utils import log_run
-from cai_lib.subprocess_utils import _run_claude_p
+from cai_lib.subagent import _run_claude_p
 
 
 # Resume-transition lookup tables. Replace the former
